Performed some refactoring to split out some of the structs into the policy and claim packages, along with the 'contructor' function for policies and claims.  NOTE: the chaincode folder must now be within your GOPATH env variable in order to build the chaincode.
